+++ src/module/data/actor/hero.mjs
@@ -4,8 +4,11 @@ import BaseEffectModel from "../effect/base.mjs";
 import { requiredInteger, setOptions } from "../helpers.mjs";
 import BaseActorModel from "./base.mjs";
 
-/** @import DrawSteelItem from "../../documents/item.mjs" */
-/** @import ActiveEffectData from "@common/documents/_types.mjs" */
+/**
+ * @import { DamageSchema } from "../pseudo-documents/power-roll-effects/_types";
+ * @import DrawSteelItem from "../../documents/item.mjs";
+ * @import ActiveEffectData from "@common/documents/_types.mjs";
+ */
 
 const fields = foundry.data.fields;
 
@@ -97,9 +100,7 @@ export default class HeroModel extends BaseActorModel {
       disengage: 0,
     };
 
-    /** @typedef {import("../item/kit.mjs").DamageSchema} DamageSchema */
-
-    this.abilityBonuses = {
+    const kitAbilityBonuses = {
       /** @type {{ distance: number, damage?: DamageSchema}} */
       melee: {
         distance: 0,
@@ -120,14 +121,16 @@ export default class HeroModel extends BaseActorModel {
       const abiBonuses = ["melee.distance", "ranged.distance"];
 
       for (const key of abiBonuses) {
-        const current = foundry.utils.getProperty(this.abilityBonuses, key);
+        const current = foundry.utils.getProperty(kitAbilityBonuses, key);
         const kitValue = foundry.utils.getProperty(bonuses, key);
-        foundry.utils.setProperty(this.abilityBonuses, key, Math.max(current, kitValue));
+        foundry.utils.setProperty(kitAbilityBonuses, key, Math.max(current, kitValue));
       }
 
-      for (const [type, obj] of Object.entries(this.abilityBonuses)) {
+      for (const [type, obj] of Object.entries(kitAbilityBonuses)) {
         if (("damage" in obj) && (this.hero.preferredKit !== kit.id)) continue;
-        if (Object.values(bonuses[type].damage).some(v => v)) obj.damage = bonuses[type].damage;
+        for (const [key, value] of Object.entries(bonuses[type].damage)) {
+          if (value) foundry.utils.setProperty(obj, `damage.${key}.value`, value);
+        }
       }
     }
 
@@ -135,6 +138,24 @@ export default class HeroModel extends BaseActorModel {
     this.movement.value += kitBonuses["speed"];
     this.combat.stability += kitBonuses["stability"];
     this.movement.disengage += kitBonuses["disengage"];
+
+    for (const keyword of ["melee", "ranged"]) {
+      const bonuses = foundry.utils.flattenObject(kitAbilityBonuses[keyword]);
+      for (const [key, value] of Object.entries(bonuses)) {
+        if (value) {
+          this._abilityBonuses.push({
+            key,
+            value,
+            mode: CONST.ACTIVE_EFFECT_MODES.ADD,
+            priority: 0,
+            filters: {
+              keywords: new Set([keyword, "weapon"]),
+            },
+          });
+        }
+
+      }
+    }
   }
 
   /* -------------------------------------------------- */
@@ -316,13 +337,6 @@ export default class HeroModel extends BaseActorModel {
    * @internal
    */
   _unfilledTraits = {};
-
-  /* -------------------------------------------------- */
-
-  /** @inheritdoc */
-  get reach() {
-    return 1 + this.abilityBonuses.melee.distance;
-  }
 
   /* -------------------------------------------------- */