+++ src/module/data/item/_types.d.ts
@@ -43,6 +43,7 @@ declare module "./ability.mjs" {
       type: keyof typeof ds["CONFIG"]["abilities"]["distances"];
       primary: number;
       secondary: number;
+      tertiary: number;
     }
     trigger: string;
     target: {
@@ -152,6 +153,11 @@ declare module "./feature.mjs" {
       value: string;
       subtype: string;
     }
+    prerequisites: {
+      value: string;
+    }
+    story: string;
+    points: number;
   }
 }